Move transaction API calls to separate file
