+++ src/hooks/useTransactions/api.js
@@ -1,0 +1,20 @@
+import ky from "ky"
+
+const fetchDomainInfo = (domain) =>
+	ky
+		.get(`/api/domain/${domain.toLowerCase()}`, {
+			throwHttpErrors: false,
+			timeout: 30000,
+		})
+		.json()
+
+const fetchTransactions = (address, before) =>
+	ky
+		.get(`/api/transactions/${address}`, {
+			searchParams: before ? { before } : "",
+			throwHttpErrors: false,
+			timeout: 30000,
+		})
+		.json()
+
+export { fetchDomainInfo, fetchTransactions }
